requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
streamlit
streamlit-lottie
kafka-python
joblib
pickle-mixin
